% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drob.R
\name{bisquare}
\alias{bisquare}
\title{Return bisquare and its derivatives}
\usage{
bisquare(k)
}
\arguments{
\item{k}{The cutoff point below and above which rho evaluates to 1.}
}
\value{
A list with three elements:
\itemize{
\item \code{rho}: The bisquare function
\item \code{psi}: The first derivative of rho
\item \code{dpsi}: The second derivative of rho
}
}
\description{
\code{bisquare} computes bisquare (aka Tukey's biweight) function for a given
cutoff point. It also computes its first two derivatives. All three functions
are returned as elements of a list, as required by the \code{step_1} parameter of
the \code{drob} function.
}
