% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drob.R
\docType{data}
\name{fpl}
\alias{fpl}
\title{The 4-parameter logistic model (4PL)}
\format{
An object of class \code{list} of length 3.
}
\usage{
fpl
}
\description{
\code{fpl} is a list containing three elements related to the 4-parameter logistic
model, as required by the \code{model} parameter of the \code{drob} function:
\itemize{
\item \code{fun}: the 4PL-function itself. It takes as arguments a vector of values
\code{x} and a vector of 4 parameters \code{t} .
\item \code{grad}: the gradient of the 4PL-function, also expressed as a function
of \code{x} and \code{t}.
\item \code{init}: a function that computes initial parameters and search bounds.
}
}
\examples{

t <- c(10, 10, 40, 20)
x <- 1:100
eta <- fpl$fun(x, t)
y <- eta + rnorm(100)
est <- fpl$init(x, y)
plot(x, y)
lines(x, eta)
lines(x, fpl$fun(x, est$t), lty = "dashed")

}
\keyword{datasets}
