% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drob.R
\name{m_scale}
\alias{m_scale}
\title{Compute M-estimate of scale}
\usage{
m_scale(r, rho, extend = 5)
}
\arguments{
\item{r}{A vector of values (typically residuals) which scale is to be
computed.}

\item{rho}{The rho function that defines the M-estimate of scale.}

\item{extend}{The interval for root finding will extend from 0 to \code{extend}
times the median absolute value of \code{r}. Defaults to 5.}
}
\value{
The \code{rho}-based M-estimate of scale for \code{r}.
}
\description{
\code{m_scale} computes an M-estimate of scale for a given rho function using
one-dimensional root finding.
}
\examples{
m_scale(r, bisquare(1.55))

}
