% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drob.R
\docType{data}
\name{fpl}
\alias{fpl}
\title{The 4-parameter logistic function (4PL)}
\format{
An object of class \code{list} of length 3.
}
\usage{
fpl
}
\description{
\code{fpl} is a list containing three elements related to the FPL model, as
required by the \code{model} parameter of the \code{drob} function:
\itemize{
\item \code{fun}: the FPL function itself. \code{fpl$fun} takes as arguments a vector
of values \code{x} and a vector of 4 parameters \code{t} .
\item \code{grad}: the gradient of the 4PL function, also expressed as a function
of \code{x} and \code{t}.
\item \code{init}: a function that computes initial parameters and search bounds.
}
}
\keyword{datasets}
